requirements.txt
fastapi
uvicorn
crewai
langchain-google-genai
pydantic
supabase
requests
python-dotenv
geopy
pandas
streamlit
pydeck
streamlit-lottie